Layout public property listings page, update some factories seeders and migrations

+++ database/migrations/2021_02_16_192606_create_property_listings_table.php
@@ -22,7 +22,7 @@ class CreatePropertyListingsTable extends Migration
             $table->unsignedInteger('bedrooms');
             $table->unsignedInteger('bathrooms');
             $table->unsignedInteger('sqft');
-            $table->decimal('rent');
+            $table->unsignedInteger('rent');
             $table->timestamps();
         });
     }
